[backend] refactored backend to separate model and views
